requirements (3).txt
deepgram-sdk
elevenlabs
websockets
anthropic
websockets
python-dotenv
pydantic
asyncpg
httpx
requests